add gender validation
