+++ src/types/common.types.ts
@@ -1,11 +1,4 @@
-export interface AppConfig {
-  port: number;
-}
-
-export interface IDictionary<T = any> {
-  [key: string]: T;
-}
-
-export interface Type<T = any> extends Function {
-  new (...args: any[]): T;
+export enum Gender {
+  MALE = 'male',
+  FEMALE = 'female',
 }
